% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDShiny.R
\name{PDShiny}
\alias{PDShiny}
\title{Launch PDShiny Interface}
\usage{
PDShiny()
}
\value{
Nothing
}
\description{
PDShiny() loads interactive user interface built using R shiny.
}
\details{
Start PDShiny

The interactive user interface is to provide an easy way for people who are learning Probability Distributions. Includes example data for testing out a few example analysis.
}
\examples{
\dontrun{
library(shiny)
PDShiny()
}
}
\keyword{PDShiny}
